Update part1a.md
We need to use the `npx create-react-app` with the `--use-npm` flag. Otherwise if the user has `yarn` installed on their system then `create-react-app` will prefer using it. So the generated app would end up with a `yarn.lock` file, instead of a `package-lock.json` file.

See https://create-react-app.dev/docs/getting-started/#selecting-a-package-manager